Split PolicyChecker from PolicyManager (#128004)
* Split PolicyChecker from PolicyManager

* Restore EntitlementCheckerUtils

* [CI] Auto commit changes from spotless

---------

(cherry picked from commit ba50798f62aade24da91efd2902f15ef9295f2fc)

# Conflicts:
#	libs/entitlement/src/main/java/org/elasticsearch/entitlement/runtime/api/ElasticsearchEntitlementChecker.java
